+++ R/subnational_alpha_cms_hat.R
@@ -1,0 +1,7 @@
+#' The median estimates of the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
+#' This array is used to inform the Normal prior of the subnational-level intercept in the single-country subnational model.
+#'
+#' @docType data
+#' @keywords subnational_alpha_cms_hat
+#' @format  A array of 23 matrices with 2 rows and 5 columns
+"subnational_alpha_cms_hat"