+++ R/tau_alphapms_subnationalmod.R
@@ -1,8 +1,0 @@
-#' tau_alphapms_subnationalmod
-#' The median estimates of the precision for the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
-#' This vector is used to inform the precision in the Normal prior of the subnational-level intercept in single-country subnational models.
-#'
-#' @docType data
-#' @keywords tau_alphapms_subnationalmod
-#' @format  A vector of two precision estimates
-"tau_alphapms_subnationalmod"