+++ NEWS.md
@@ -1,3 +1,15 @@
 # mcmsupply 0.2.0
 
 * Initial CRAN submission.
+
+# mcmsupply 0.2.1
+
+* Updates following reviewers comments on the national supply share model
+* Improved standard error imputation for missing observations
+* Removal of truncated data sets from package - no longer required.
+
+# mcmsupply 0.2.2
+
+* Updates to data for new single-country parameters
+* Updates to data for multi-country correlations 
+* Updates for Custom data bug fix